+++ leda/components/Tooltip/constants.ts
@@ -1,3 +1,10 @@
 export const defaultArrowSize = 7;
 export const defaultPosition = 'top';
 export const defaultTransitionTimeout = 200;
+
+export enum Position {
+  Top = 'top',
+  Bottom = 'bottom',
+  Right = 'right',
+  Left = 'left',
+}