+++ leda/components/Tour/types.ts
@@ -1,15 +1,45 @@
 import * as React from 'react';
 
+/**
+ * Change event
+ */
+export interface ChangeEvent {
+  component: {
+    /** Current step key */
+    value: number | string | null,
+    /** Current TourStepItem */
+    item: TourStepItem | null,
+  },
+}
+
+/**
+ * Modal window content properties
+ */
 export interface ContentProps {
+  /** Function allows to jump to any tour item */
   goTo: (stepNumber: number) => void,
+  /** Function switches to next tour item */
   next: () => void,
+  /** Function switches to previous tour item */
   prev: () => void,
+  /** Function stops tour  */
   stopTour: () => void,
+}
+
+export interface TourProps {
+  /** Идентификатор активного шага, если null - гайд-тур закрыт */
+  activeStepKey?: number | string | null,
+  /** Шаги гайд-тура */
+  data: TourStepItem[],
+  /** Обработчик изменения */
+  onChange: (ev: ChangeEvent) => void,
 }
 
 export interface TourStepItem {
   /** Скругление у выделенных элементов в px, по-умолчанию 15px */
   borderRadius?: number,
+  /** Отступы выделенной области в px, по-умолчанию 0px */
+  padding?: number,
   /** Контент */
   content: (props: ContentProps) => React.ReactElement | null,
   /** Элемент, который надо подсветить */
@@ -20,20 +50,4 @@ export interface TourStepItem {
   position: 'top' | 'right' | 'bottom' | 'left',
   /** Идентификатор шага */
   stepKey: string | number,
-}
-
-export interface ChangeEvent {
-  component: {
-    value: number | string | null,
-    item: TourStepItem | null,
-  },
-}
-
-export interface TourProps {
-  /** Идентификатор активного шага, если null - гайд-тур закрыт */
-  activeStepKey?: number | string | null,
-  /** Шаги гайд-тура */
-  data: TourStepItem[],
-  /** Обработчик изменения */
-  onChange: (ev: ChangeEvent) => void,
 }