+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.34.0] - 2020-11-09
+
+### Added
+- Set up automatic calculation of test coverage
+- Tour: Added padding prop and possibility to interact with tour element
+
+### Fixed
+- DatePicker, TimePicker: Fixed validation
+- MultiSelect: Input is cleared after selecting item
+- ToolTip: Added corner position
+- Validation: Fixed validation in controlled mode
+
 ## [0.33.0] - 2020-10-26
 
 ### Added